CUDA port: Moves checkCudaErrors to separate header
Simplifies the compliation guards needed for the non-cuda source files
